Update security, calendar, and ordering related files

+++ tlqkf/tlqkf/src/main/java/com/ontherocks/tlqkf/controller/CalendarController.java
@@ -3,6 +3,8 @@ package com.ontherocks.tlqkf.controller;
 import com.ontherocks.tlqkf.service.CalendarService;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/calendar")
 public class CalendarController {
@@ -13,7 +15,7 @@ public class CalendarController {
     }
 
     @GetMapping("/sales")
-    public int getTotalSales() {
-        return calendarService.calculateTotalSales();
+    public BigDecimal getTotalSales() {
+        return calendarService.calculateTotalSales(); // ✅ 반환 타입을 `BigDecimal`로 변경
     }
 }
